refactor: move decorators next to their consumers
- Move @needs_lock to execution/decorators.py (near the lock interceptor)
- Move @mcp_tool to server/mcp/decorators.py (near the MCP server)
- Delete application_sdk/decorators/ directory entirely
- Re-export needs_lock from execution/__init__.py for clean public API
- Update all test imports and docs

Closes: BLDX-880
