+++ application_sdk/execution/__init__.py
@@ -7,10 +7,12 @@ from application_sdk.execution._temporal.auth import (
 from application_sdk.execution._temporal.backend import create_temporal_client
 from application_sdk.execution._temporal.converter import create_data_converter
 from application_sdk.execution._temporal.worker import AppWorker, create_worker
+from application_sdk.execution.decorators import needs_lock
 from application_sdk.execution.errors import ApplicationError
 
 __all__ = [
     "ApplicationError",
+    "needs_lock",
     "AppWorker",
     "create_worker",
     "create_temporal_client",
